renamed hibernatedemo to hibernate4 with modifications to readme and improved comments.
