+++ hibernate4/src/main/java/org/jboss/as/quickstart/hibernate4/util/Resources.java
@@ -1,4 +1,4 @@
-package org.jboss.as.quickstart.hibernatedemo.util;
+package org.jboss.as.quickstart.hibernate4.util;
 
 import java.util.logging.Logger;
