+++ hibernate4/src/main/resources/META-INF/persistence.xml
