+++ hibernate4/src/main/webapp/WEB-INF/beans.xml
