+++ hibernate4/src/main/webapp/WEB-INF/faces-config.xml
