+++ hibernate4/src/main/webapp/WEB-INF/templates/default.xhtml
