+++ hibernate4/src/main/webapp/index.xhtml
