+++ hibernate4/src/main/webapp/resources/gfx/logo.png
