--- hibernatedemo/src/main/resources/import.sql
@@ -1,3 +1,0 @@
--- You can use this file to load seed data into the database using SQL statements
-insert into Member (id, name, email, phone_number, address) values (0, 'John Smith', 'john.smith@mailinator.com', '2125551212', 'Boston NY') 
-insert into Member (id, name, email, phone_number, address) values (1, 'Madhumita Sadhukhan', 'msadhukh@gmail.com', '2135551214', 'Brno CZ')
